requirements.txt
crewai
crewai-tools
requests
beautifulsoup4
python-dotenv
openai
langchain